Modify FDSN logic to acknowledge multiple channels
There was a hardcoded assumption in get_inventory_from_df that
only one valid value of channel instance should be added to the list
of returned channels from a station.

This is not only false in the case of wildcards (e.g. *F*, *Q*),
it is also false when a channel has been broken into multiple runs.

CAS04 is a good example of this.

See mth5 issue #157, and aurora 277 for more details.