hotfix synthetic electric field polarity (#236)
A synthetic e-field flip correction was recently added
This reversed the polarity of the  'ex', 'ey' columns of dataframe
right after it gets read-in.
However, the dataframe column names come from the channel nomenclature,
and these can be, for example 'e1', 'e2'.

Some notes about a permanent fix are in #236